backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Vector database
qdrant-client==1.7.0

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2

# HTTP client
httpx==0.26.0

# Utilities
numpy==1.26.3
